+++ LoopKitUI/Extensions/DeviceStatusHighlight+Color.swift
@@ -1,26 +1,12 @@
 //
-//  DeviceLifecycleProgressState+Color.swift
+//  DeviceStatusHighlight+Color.swift
 //  LoopKitUI
 //
-//  Created by Nathaniel Hamming on 2020-07-03.
+//  Created by Nathaniel Hamming on 2020-07-06.
 //  Copyright © 2020 LoopKit Authors. All rights reserved.
 //
 
 import LoopKit
-
-extension DeviceLifecycleProgressState {
-    public var color: UIColor {
-        switch self {
-        case .normal:
-            return .systemPurple
-        case .warning:
-            return .systemOrange
-        case .critical:
-            return .systemRed
-        }
-    }
-}
-
 
 extension DeviceStatusHighlight {
     public var image: UIImage {
